feat(1127): utilisation du même formulaire dans la modale et dans la page profil
